% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_airtemp.R
\name{get_airtemp}
\alias{get_airtemp}
\title{Gets per-minute air temperature readings in Singapore for specified set of dates.}
\usage{
get_airtemp(from, to)
}
\arguments{
\item{from}{Starting date. Must be a character string, in "YYYY-MM-DD" format.}

\item{to}{Final date, inclusive. Must be a character string, in "YYYY-MM-DD" format.}
}
\value{
Per-minute air temperature readings in Singapore for various weather stations.
}
\description{
Returns a dataframe of per-minute air temperature readings for specified dates for various weather stations in Singapore.
}
\details{
User inputs a starting and final date. Per-minute air temperature readings are then given for all dates from the starting date to the final date, inclusive. Temperature is in degrees Celsius. Data is provided for up to 16 weather stations across Singapore, depending on availability of data. To obtain location of weather stations, use function get_weatherstns.
}
\examples{
get_airtemp("2017-12-31", "2018-01-04")
}
