% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_pollutants.R
\name{get_pollutants}
\alias{get_pollutants}
\title{Gets pollutant levels in Singapore for specified set of dates.}
\usage{
get_pollutants(from, to, natl_only = T)
}
\arguments{
\item{from}{Starting date. Must be a character string, in "YYYY-MM-DD" format.}

\item{to}{Final date, inclusive. Must be a character string, in "YYYY-MM-DD" format.}

\item{natl_only}{True or False, defaults to true. If true, returns only national values. If false, returns data for national and 5 subnational regions.}
}
\value{
Pollutant indices readings in Singapore for national and subnational regions for selected dates.
}
\description{
Returns a dataframe of pollutant levels for dates between starting and final date for national and 5 subnational regions.
}
\details{
User inputs a starting and final date. Hourly pollutant levels are then given for all dates from the starting date to the final date, inclusive. Data is provided for national and 5 subnational regions, in the form of 12 pollutant indices: O3 Sub-Index, PM10 24-hourly, PM10 Sub-Index, CO Sub-Index, PM2.5 24-Hourly, SO2 Sub-Index, CO 8-Hour-Max, NO2 1-Hour-Max, SO2 24-Hourly, PM2.5 Sub-Index, PSI 24-Hourly and O3 8-Hour Max
}
\examples{
get_pollutants("2017-12-31", "2018-01-04")
#' get_pollutants("2018-07-05", "2018-08-06", natl_only = F)
}
