% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_rainfall.R
\name{get_rainfall}
\alias{get_rainfall}
\title{Gets 5-minute rainfall readings in Singapore for specified set of dates.}
\usage{
get_rainfall(from, to)
}
\arguments{
\item{from}{Starting date. Must be a character string, in "YYYY-MM-DD" format.}

\item{to}{Final date, inclusive. Must be a character string, in "YYYY-MM-DD" format.}
}
\value{
5-minute rainfall readings in Singapore for various weather stations.
}
\description{
Returns a dataframe of 5-minute rainfall readings for specified dates for various weather stations in Singapore.
}
\details{
User inputs a starting and final date. 5-minute rainfall readings are then given for all dates from the starting date to the final date, inclusive. Rainfall is measured by mm/5 min. Data is provided for up to 52 weather stations across Singapore, depending on availability of data. To obtain location of weather stations, use function get_weatherstns.
}
\examples{
get_rainfall("2017-12-31", "2018-01-04")
}
