% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_UV.R
\name{get_UV}
\alias{get_UV}
\title{Gets UV levels in Singapore for specified set of dates.}
\usage{
get_UV(from, to)
}
\arguments{
\item{from}{Starting date. Must be a character string, in "YYYY-MM-DD" format.}

\item{to}{Final date, inclusive. Must be a character string, in "YYYY-MM-DD" format.}
}
\value{
UV levels in Singapore for dates between starting and final date.
}
\description{
Returns a dataframe of UV levels for dates between starting and final date.
}
\details{
User inputs a starting and final date. UV levels are then given for all dates from the starting date to the final date, inclusive. Data is available for when the sun is up - from 0700 to 1900hrs.
}
\examples{
get_UV("2017-12-31", "2018-01-04")
}
