% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_weatherstns.R
\name{get_weatherstns}
\alias{get_weatherstns}
\title{Gets data on weather stations in Singapore.}
\usage{
get_weatherstns()
}
\value{
Dataframe with weather station IDs, locations and latitude-longitude coordinates
}
\description{
Returns a dataframe of weather station IDs, names and coordinates.
}
\examples{
get_weatherstns()
}
