% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_weathersumm.R
\name{get_weathersumm}
\alias{get_weathersumm}
\title{Gets hourly aggregated weather data in Singapore for specified set of dates.}
\usage{
get_weathersumm(from, to)
}
\arguments{
\item{from}{Starting date. Must be a character string, in "YYYY-MM-DD" format.}

\item{to}{Final date, inclusive. Must be a character string, in "YYYY-MM-DD" format.}
}
\value{
Dataframe of air temperature, humidity, rainfall, UV, PSI 24-hourly data aggregated over Singapore at the hourly level.
}
\description{
Returns a dataframe of hourly aggregated weather data in Singapore for specified dates: Air temperature, Humidity, Rainfall, UV, PSI 24-Hourly
}
\details{
User inputs a starting and final date. Returns hourly aggregated weather data from multiple API sources. Per-minute air temperature data, per-minute humidity data, 5-minute rainfall data is aggregated over the hour and across weather stations distributed across Singapore. PSI 24-hourly data and UV levels will be taken directly from APIs which provide national, hourly data.
}
\examples{
get_weathersumm("2017-12-31", "2018-01-04")
}
